requirements.txt
# Web Framework
streamlit==1.32.0

# AI / LLM
google-generativeai==0.5.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2
xlrd==2.0.1

# Database
sqlalchemy==2.0.27

# Visualization
plotly==5.19.0

# Environment Variables
python-dotenv==1.0.1

# Utilities
Pillow==10.2.0
tabulate==0.9.0

# Export Features
fpdf2==2.7.9
xlsxwriter==3.1.9